% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sra_db.R
\name{download_sra_db}
\alias{download_sra_db}
\title{Download SRA metadata database}
\usage{
download_sra_db(file.name = "SRAmetadb.sqlite", path = getwd(),
  method = "auto")
}
\arguments{
\item{file.name}{File name for uncompressed database. Default: "SRAmetadb.sqlite"}

\item{path}{Path used to save database. If not provided, current directory is used.}

\item{method}{Method for downloading database: "auto", "internal", "libcurl", "wget", "curl", "lynx". See
\code{\link{download.file}} function.}
}
\description{
Downloads and extracts SRA metadata database
}

