% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fastq_urls.R
\name{get_fastq_urls}
\alias{get_fastq_urls}
\title{Get urls for downloading fastq files}
\usage{
get_fastq_urls(runacc, db_conn)
}
\arguments{
\item{runacc}{A vector with run accessions}
}
\value{
urls A nested list with run accessions separated depending on whether they are associated with one, two or three files.
If all run accessions provided are only associated with one category, a data frame is returned for that 
category.
}
\description{
Gets all urls, file names and md5 values associated with run accessions.
}
\examples{
library('srametadata')

# Get path to sqlite database
sqlfile <- file.path('.', 'SRAmetadb.sqlite')

# Create connection
conn <- dbConnect(SQLite(),sqlfile)

# one file
get_fastq_urls(runacc = c("DRR001482"), db_con = conn)

# two files
get_fastq_urls(runacc = c("DRR001627"),  db_con = conn)

# three files
get_fastq_urls(runacc = c("ERR204978"),  db_con = conn)

# one file and two files
get_fastq_urls(runacc = c("DRR001482", "DRR001627"),  db_con = conn)

# one file and three files
get_fastq_urls(runacc = c("DRR001482", "ERR204978"),  db_con = conn)

# one file and three files
get_fastq_urls(runacc = c("DRR001627", "ERR204978"),  db_con = conn)

# one file, two files and three files
get_fastq_urls(runacc = c("DRR001482", "DRR001627", "ERR204978"),  
               db_con = conn)
}

