% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_metadata.R
\name{get_raw_metadata}
\alias{get_raw_metadata}
\title{Get all raw metadata from SRA}
\usage{
get_raw_metadata(run_accession = NULL, sample_accession = NULL,
  experiment_accession = NULL, study_accession = NULL,
  submission_accession = NULL, platform = NULL, db_conn)
}
\arguments{
\item{run_accession}{A vector with run accessions}

\item{sample_accession}{A vector with sample accessions}

\item{experiment_accession}{A vector with experiment accessions}

\item{study_accession}{A vector with study accessions}

\item{submission_accession}{A vector with submission accessions}

\item{platform}{Sequencing platform, e.g "ILLUMINA"}

\item{db_connn}{Connection to the sqlite database}
}
\value{
A dataframe with all metadata associated with the filters provided to 
the function.
}
\description{
Get all raw metadata from SRA associated with SRA accessions numbers
}
\examples{
library('srametadata')

sqlfile <- file.path('.', 'SRAmetadb.sqlite')

# Create connection
conn <- dbConnect(SQLite(),sqlfile)

get_raw_metadata(run_accession = c("DRR023427", "SRR946066", "ERR204978"), 
                 experiment_accession = "SRX329581", 
                 sample_accession = c("SRS465599","DRS014276"),
                 study_accession = "SRP028344",
                 submission_accession = "SRA096347", 
                 platform = "ILLUMINA",
                 db_conn = conn)

get_raw_metadata(experiment_accession = "SRX329581", 
                 db_conn = conn)

get_raw_metadata(
    run_accession = c("DRR023427", "SRR946066", "ERR204978"), 
    db_conn = conn)

get_raw_metadata(
    sample_accession = c("SRS465599","DRS014276"),
    db_conn = conn)

all_illumina_sra <- get_raw_metadata(platform = "ILLUMINA", db_conn = conn)
}

