% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srametaintrons.R
\name{get_srameta_introns}
\alias{get_srameta_introns}
\title{merges srametadata with intron counts}
\usage{
get_srameta_introns(file1, file2, file3)
}
\arguments{
\item{file1}{file1 is a file containing intron counts on chrY for sra metadata database}

\item{file2}{file2 is a file containing intron counts on whole chromosomes for sra metadata database}

\item{file3}{file3 is a file containing all run accessions for sra metadata database}
}
\value{
return a list of dataframes: srametaintrons.Rda, srametaintrons_subset.Rda
}
\description{
saves srametaintrons.Rda with 20,172 entries
saves srametaintrons_subset.Rda with 2,969 entries 
containing complete sex records from SRA metadata database
}
\examples{
get_srameta_introns("chrYintrons.txt","totalintrons.txt","run_accession.txt")
}

