% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genenormalization.R
\name{normalized}
\alias{normalized}
\title{normalize gene expressions for sra metadata database}
\usage{
normalized(file, path, chr, start, end)
}
\arguments{
\item{file}{file:filename}

\item{path}{path used in the .sh file}

\item{chr}{chr specify which chromosome is inspected}

\item{start}{start specify the starting position of the inspected gene}

\item{end}{end specify the ending position of the inspected gene}
}
\value{
a normalized gene expression dataframe
}
\description{
normalize gene expressions for sra metadata database
}
\examples{
library(rtracklayer)
library(GenomicRanges)
normalized(file = scan("sra_samples.txt", what="", sep="\\n"), path = path, chr = "chrY", start = 14813160, end = 14972768)
}

