% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexprediction.R
\name{predictsex}
\alias{predictsex}
\title{predict sex information based on intron counts for SRA metadata database}
\usage{
predictsex(run_accession = run_accession)
}
\arguments{
\item{run_accession}{A vector with run accessions}
}
\value{
predicted sex and prediction accuracy to the corresponding run accession input
}
\description{
predict sex information based on intron counts for SRA metadata database
}
\examples{
predictsex(run_accession = "SRR1296065")
predictsex(run_accession = c("SRR1296065","SRR1044429"))
}

