% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_field.R
\name{search_field}
\alias{search_field}
\title{Get metadata from SRA}
\usage{
search_field(runacc, field, type, db_conn)
}
\arguments{
\item{runacc}{A vector with run accessions}

\item{field}{A string specifying the field of interest (e.g. "sex", "cell_line", "tissue")}

\item{type}{Type of metadata which is associated with runs: 'sample', 'experiment', 'study' or 'run'}

\item{db_conn}{Connection to the sqlite database}
}
\value{
\code{resfield} A dataframe with run_accessions and values associated with the field of interest. 
If field does not exist, NAs are returned
}
\description{
This function grabs metadata from the sample, experiment, study and experiment attribute in the
SRA metadata database.
}
\examples{
search_field(runacc = c("DRR023427", "SRR946066", "ERR204978"),  
           field = "disease", 
           type = "sample",
           db_con = conn)

}

